WIP: Almost there with getting redux to work
The following commit includes some tweaks to get redux working. It's still
a work in progress however. With the current set up, I'm running into a infinite
loop in the browser... Thus far the source of that infinite loop isn't totally clear.
Forcing useEffect to return an empty array is the current bandaid in place to
prevent that infinite loop from happening.